Update Dockerfile Install OhMyZsh to shell
This commit installs OhMyZsh early in the Dockerfile in the challenge folder so when a challenge is loaded and whole the vm is being set up ill install OhMyZsh. This will provide Zsh as the default shell for improved user experience and to enable subsequent configurations and tools to use Zsh.